Extract content/storage helpers into lib + more fuzzing
