+++ index.js
@@ -4,10 +4,8 @@ const { EventEmitter } = require('events')
 const collect = require('stream-collector')
 const thunky = require('thunky')
 const unixify = require('unixify')
-const raf = require('random-access-file')
 const mutexify = require('mutexify')
 const duplexify = require('duplexify')
-const sodium = require('sodium-universal')
 const through = require('through2')
 const pump = require('pump')
 
@@ -18,6 +16,8 @@ const coreByteStream = require('hypercore-byte-stream')
 const Stat = require('./lib/stat')
 const errors = require('./lib/errors')
 const messages = require('./lib/messages')
+const { defaultStorage } = require('./lib/storage')
+const { contentKeyPair, contentOptions } = require('./lib/content')
 
 class Hyperdrive extends EventEmitter {
   constructor (storage, key, opts) {
@@ -530,63 +530,6 @@ module.exports = Hyperdrive
 
 function isObject (val) {
   return !!val && typeof val !== 'string' && !Buffer.isBuffer(val)
-}
-
-function wrap (self, storage) {
-  return {
-    metadata: function (name, opts) {
-      return storage.metadata(name, opts, self)
-    },
-    content: function (name, opts) {
-      return storage.content(name, opts, self)
-    }
-  }
-}
-
-function defaultStorage (self, storage, opts) {
-  var folder = ''
-
-  if (typeof storage === 'object' && storage) return wrap(self, storage)
-
-  if (typeof storage === 'string') {
-    folder = storage
-    storage = raf
-  }
-
-  return {
-    metadata: function (name) {
-      return storage(path.join(folder, 'metadata', name))
-    },
-    content: function (name) {
-      return storage(path.join(folder, 'content', name))
-    }
-  }
-}
-
-function contentOptions (self, secretKey) {
-  return {
-    sparse: self.sparse || self.latest,
-    maxRequests: self.maxRequests,
-    secretKey: secretKey,
-    storeSecretKey: false,
-    indexing: self.metadataFeed.writable && self.indexing,
-    storageCacheSize: self.contentStorageCacheSize
-  }
-}
-
-function contentKeyPair (secretKey) {
-  var seed = Buffer.allocUnsafe(sodium.crypto_sign_SEEDBYTES)
-  var context = Buffer.from('hyperdri', 'utf8') // 8 byte context
-  var keyPair = {
-    publicKey: Buffer.allocUnsafe(sodium.crypto_sign_PUBLICKEYBYTES),
-    secretKey: Buffer.allocUnsafe(sodium.crypto_sign_SECRETKEYBYTES)
-  }
-
-  sodium.crypto_kdf_derive_from_key(seed, 1, context, secretKey)
-  sodium.crypto_sign_seed_keypair(keyPair.publicKey, keyPair.secretKey, seed)
-  if (seed.fill) seed.fill(0)
-
-  return keyPair
 }
 
 function split (buf) {
