+++ lib/storage.js
@@ -1,0 +1,38 @@
+const path = require('path')
+
+const raf = require('random-access-file')
+
+function wrap (self, storage) {
+  return {
+    metadata: function (name, opts) {
+      return storage.metadata(name, opts, self)
+    },
+    content: function (name, opts) {
+      return storage.content(name, opts, self)
+    }
+  }
+}
+
+function defaultStorage (self, storage, opts) {
+  var folder = ''
+
+  if (typeof storage === 'object' && storage) return wrap(self, storage)
+
+  if (typeof storage === 'string') {
+    folder = storage
+    storage = raf
+  }
+
+  return {
+    metadata: function (name) {
+      return storage(path.join(folder, 'metadata', name))
+    },
+    content: function (name) {
+      return storage(path.join(folder, 'content', name))
+    }
+  }
+}
+
+module.exports = {
+  defaultStorage
+}
